Define a macro PRINTF_F in dd_share.h, which indicates printf-like functions, so that GCC can check their parameters. If GCC is not the compiler, PRINTF_F defines to nothing.
Make sure the parameters match the format string for all printf-like functions in glBSP.

Thanks to chasonr for the patch.

+++ doomsday/external/glbsp/src/system.c
@@ -251,16 +251,16 @@ void InitEndian(void)
   /* sanity-check type sizes */
 
   if (sizeof(uint8_g) != 1)
-    FatalError("Sanity check failed: sizeof(uint8_g) = %d", 
-        sizeof(uint8_g));
+    FatalError("Sanity check failed: sizeof(uint8_g) = %lu", 
+               (unsigned long) sizeof(uint8_g));
 
   if (sizeof(uint16_g) != 2)
-    FatalError("Sanity check failed: sizeof(uint16_g) = %d", 
-        sizeof(uint16_g));
+    FatalError("Sanity check failed: sizeof(uint16_g) = %lu", 
+               (unsigned long) sizeof(uint16_g));
 
   if (sizeof(uint32_g) != 4)
-    FatalError("Sanity check failed: sizeof(uint32_g) = %d", 
-        sizeof(uint32_g));
+    FatalError("Sanity check failed: sizeof(uint32_g) = %lu", 
+               (unsigned long) sizeof(uint32_g));
 
   /* check endianness */
